LayoutTests/imported/w3c/web-platform-tests/wasm/webapi/esm-integration/wasm-to-wasm-link-error.tentative-expected.txt

CONSOLE MESSAGE: LinkError: imported global ./wasm-export-i64-global.wasm:g must be a same type

PASS Link errors for imports between WebAssembly modules should be reported.
